Move/rename tests and scss files
